refactor: split CLIFacade into domain extension files (#428)
CLIFacade.swift (1,746 → 860 lines) now contains struct definition,
shared helpers, and types. Method groups extracted to:
- CLIFacade+Rules.swift — custom rules CRUD
- CLIFacade+Collections.swift — rule collections CRUD
- CLIFacade+Service.swift — service lifecycle, config, TCP, status
- CLIFacade+Packs.swift — pack management

No API changes — all commands continue to use CLIFacade unchanged.
